requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.3
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
decorator==5.1.1
discord.py==2.3.1
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
multidict==6.0.4
mutagen==1.46.0
pefile==2023.2.7
pycparser==2.21
pycryptodomex==3.18.0
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.4
PyNaCl==1.5.0
pywin32-ctypes==0.2.2
requests==2.31.0
sniffio==1.3.0
urllib3==2.0.3
validators==0.20.0
websockets==11.0.3
yarl==1.9.2
youtube-dl==2021.6.6
youtube-search==2.1.2
youtube-search-python==1.6.6
